# dakunlun/docs/swagger.yaml
definitions:
...
  constant.Paging:
    properties:
      page:
        type: integer
      perPage:
        type: integer
      totalNum:
        type: integer
    type: object
...
  msg.Mail:
    properties:
      args:
        description: 占位符参数
        items:
          type: string
        type: array
      attachment:
        description: 附件（通用奖励格式）
        items:
          type: string
        type: array
      createTime:
        description: 创建时间
        type: integer
      hasReceived:
        description: 是否领取附件
        type: integer
      id:
        description: 数据库ID
        type: integer
      mailID:
        description: 邮件模板ID
        type: integer
      status:
        description: 状态 0未读 1已读 2已删除
        type: integer
    type: object
  msg.MailListResponse:
    properties:
      mails:
        items:
          $ref: '#/definitions/msg.Mail'
        type: array
      pageInfo:
        $ref: '#/definitions/constant.Paging'
    type: object
  msg.MailReadResponse:
    type: object
  msg.MailReceiveAllResponse:
    properties:
      rewards:
        items:
          $ref: '#/definitions/msg.Reward'
        type: array
    type: object
  msg.MailReceiveResponse:
    properties:
      rewards:
        items:
          $ref: '#/definitions/msg.Reward'
        type: array
    type: object
...
  msg.Reward:
    properties:
      mainType:
        type: integer
      subType:
        type: integer
      val:
        type: integer
    type: object
...
  msg.Tech:
    properties:
      id:
        type: integer
      level:
        type: integer
      techID:
        type: integer
    type: object
  msg.TechListResponse:
    properties:
      techs:
        items:
          $ref: '#/definitions/msg.Tech'
        type: array
    type: object
  msg.TechUpgradeResponse:
    properties:
      level:
        type: integer
    type: object
...
host: 8.140.113.88:8080
info:
  contact: {}
  description: 接口文档
  title: dakunlun
  version: "1.0"
paths:
...
  /api/mail/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 页号
        in: query
        name: page
        required: true
        type: integer
      - description: 页号
        in: query
        name: perPage
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.MailListResponse'
      summary: 邮件列表
  /api/mail/read:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 唯一ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.MailReadResponse'
      summary: 标记已读
  /api/mail/receive:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 唯一ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.MailReceiveResponse'
      summary: 领取附件
  /api/mail/receiveall:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.MailReceiveAllResponse'
      summary: 邮件批量领取
  /api/tech/list:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.TechListResponse'
      summary: 科技列表
  /api/tech/upgrade:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户ID
        in: header
        name: GameUid
        required: true
        type: string
      - description: 令牌
        in: header
        name: GameToken
        required: true
        type: string
      - description: 请求序号
        in: header
        name: SeqId
        required: true
        type: integer
      - description: 科技自增ID
        in: query
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/msg.TechUpgradeResponse'
      summary: 科技升级
...
swagger: "2.0"